requirements.txt
alembic==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
backoff==2.2.1
bcrypt==5.0.0
black==24.10.0
boto3==1.35.95
botocore==1.35.95
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.115.6
filelock==3.24.3
flake8==7.1.1
flatbuffers==25.12.19
fsspec==2026.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
jmespath==1.1.0
kubernetes==35.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
posthog==7.9.4
protobuf==6.33.5
pyasn1==0.6.2
pycodestyle==2.12.1
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==5.5.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2023.3
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
s3transfer==0.10.4
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.20.3
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xlrd==2.0.1
zipp==3.23.0
